move function, tidy up
